airline-demand-app/core/requirements.txt
streamlit
pandas
requests
python-dotenv
plotly
opensky-api
